Examine Cyfral+Metakom protocol [READ DESC]
Explaining the formats. Also attempting to add readability for widespread copiers like TMD.

EM_Marin readability on this is for another day... All you have to know is that the FF in the beginning is actually not necessary, so there's that, but I still dunno how to get to regular 5 byte codes.